JSONResponse on storage.js somehow fiddles with RESPONSE.data
As a result, if the data of the response itself contains a field "data",
this one is used as "data" component of the  JSONResponse. Otherwise
response is itself wrapped into this.data. So we have

this.data = response.data if that is set, otherwise
this.data = response

Some functions in contacts.js did not follow this scheme.